pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "source code for the Red Line Simulation project"
authors = ["Yangxinyu <xinyux@wharton.upenn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
selenium = "^4.15.2"
gtfsutils = "^0.0.5"
mapclassify = "2.6.0"
attrs = "23.1.0"
branca = "0.6.0"
census = "0.8.20"
certifi = "2023.7.22"
charset-normalizer = "3.2.0"
click-plugins = "1.1.1"
click = "8.1.7"
cligj = "0.7.2"
colorama = "0.4.6"
configargparse = "1.7"
contourpy = "1.1.0"
cycler = "0.11.0"
filelock = "3.12.3"
fiona = "1.9.4.post1"
folium = "0.14.0"
fonttools = "4.42.1"
geographiclib = "2.0"
geopandas = "0.13.2"
geopy = "2.4.0"
idna = "3.4"
importlib-resources = "6.0.1"
jellyfish = "0.11.2"
jinja2 = "3.1.2"
joblib = "1.3.2"
jpype1 = "1.4.1"
kiwisolver = "1.4.5"
markupsafe = "2.1.3"
matplotlib = "3.7.2"
networkx = "3.1"
numpy = "1.25.2"
packaging = "23.1"
pandas = "2.0.3"
pillow = "10.0.0"
psutil = "5.9.5"
pyparsing = "3.0.9"
pyproj = "3.6.0"
python-dateutil = "2.8.2"
pytz = "2023.3"
pyyaml = "6.0.1"
r5py = "0.1.0"
requests = "2.31.0"
scikit-learn = "1.3.0"
scipy = "1.9.3"
shapely = "2.0.1"
six = "1.16.0"
threadpoolctl = "3.2.0"
tqdm = "4.66.1"
typing-extensions = "4.7.1"
tzdata = "2023.3"
urllib3 = "2.0.4"
us = "3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
